unit tests should not have a need on external servies like kind.  refactored to move from unit to integration
