+++ pkg/test/harness_integration_test.go
@@ -3,7 +3,10 @@
 package test
 
 import (
+	"syscall"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 
 	harness "github.com/kudobuilder/kuttl/pkg/apis/testharness/v1beta1"
 )
@@ -19,4 +22,29 @@ func TestHarnessRunIntegration(t *testing.T) {
 		T: t,
 	}
 	harness.Run()
+}
+
+// This test requires external KinD support to run thus is an integration test
+func TestRunBackgroundCommands(t *testing.T) {
+	h := Harness{
+		T: t,
+	}
+	h.TestSuite.StartControlPlane = true
+	commands := []harness.Command{{
+		Command:    "sleep 1000000",
+		Background: true,
+	}}
+	h.TestSuite.Commands = commands
+
+	h.Setup()
+	defer h.Stop()
+
+	// setup creates bg processes
+	assert.Equal(t, 1, len(h.bgProcesses))
+	// process is alive
+	assert.NoError(t, h.bgProcesses[0].Process.Signal(syscall.Signal(0)))
+
+	// cleans up bg processes
+	h.Stop()
+	assert.Error(t, h.bgProcesses[0].Process.Signal(syscall.Signal(0)))
 }
